Network.API.PagerDuty.EventV1

Description

PagerDuty Event V1 interface.

Triggering an Event

data TriggerEvent a

Request object to create an event. Any value that may be serialized to JSON maybe attached as details.

This may be delivered with the deliver function.

Constructors

TriggerEvent

Fields

_teServiceKey :: Text

_teIncidentKey :: Maybe Text

_teDescription :: Text

_teDetails :: Maybe a

_teClient :: Text

_teClientURL :: Text

_teContexts :: [Context]

type TriggerEvent' = TriggerEvent ()

A TriggerEvent type that doesn't have details.

data Context

Context that may be added when creating an event.

Constructors

Link Text (Maybe Text)
Link to a URL with an optional link description.

Image Text (Maybe Text) (Maybe Text)
Image URL, optional link ref, and optional alt text.

Updating an Event

data UpdateEvent a

UpdateEvent is the message for both acknowledging and resolving incidents. This may be delivered using the deliver function.

Constructors

UpdateEvent

Fields

_updateType :: UpdateType

_updateServiceKey :: Text

_updateIncidentKey :: Text

_updateDescription :: Text

_updateDetails :: Maybe a

type UpdateEvent' = UpdateEvent ()

A UpdateEvent type that doesn't have details.

data UpdateType

An event update will either acknowledge or resolve an incident.

Constructors

Acknowledge

Resolve

Delivering Events to PagerDuty

deliver :: (EventRequest r, MonadCatch m, MonadIO m) => r -> m Response

Deliver a TriggerEvent or UpdateEvent.

data Response

Response to a delivered message.

Constructors

Failure Text Text
Failure status and message

Success Text
Success and incident key for further updates

Lenses and Prisms

_Link :: Prism' Context (Text, Maybe Text)

_Image :: Prism' Context (Text, Maybe Text, Maybe Text)

teServiceKey :: forall a. Lens' (TriggerEvent a) Text

teIncidentKey :: forall a. Lens' (TriggerEvent a) (Maybe Text)

teDescription :: forall a. Lens' (TriggerEvent a) Text

teDetails :: forall a a. Lens (TriggerEvent a) (TriggerEvent a) (Maybe a) (Maybe a)

teClient :: forall a. Lens' (TriggerEvent a) Text

teClientURL :: forall a. Lens' (TriggerEvent a) Text

teContexts :: forall a. Lens' (TriggerEvent a) [Context]

_Acknowledge :: Prism' UpdateType ()

_Resolve :: Prism' UpdateType ()

updateType :: forall a. Lens' (UpdateEvent a) UpdateType

updateServiceKey :: forall a. Lens' (UpdateEvent a) Text

updateIncidentKey :: forall a. Lens' (UpdateEvent a) Text

updateDescription :: forall a. Lens' (UpdateEvent a) Text

updateDetails :: forall a a. Lens (UpdateEvent a) (UpdateEvent a) (Maybe a) (Maybe a)